Fix emmalloc_trim() to work again (#25126)
Found problems:

1. The convention for sbrk() changed at some point to require 16b
alignment. (`__alignof__(max_align_t)`) This can create gaps in the
alignment of emmalloc heap end sentinel region and the previous free
region. This must require changing the assumption in emmalloc that heap
end sentinel region would always be fixed 16b -> allow it to be
arbitrary sized (in practice 16-31b).
2. sbrk() implementation had round-up to nearest max_align_t, but didn't
handle negative sbrk() values.
3. in 4GB mode, a 32-bit sbrk() is not enough to accommodate support for
both sbrk() increases and decreases (unless we take a convention that
sbrk() can only be increased by < 2GB amount.. which probably won't hold
in use cases that want to claim the whole heap immediately). Add a
function sbrk64() that can be used in 4GB builds to be able to express
both memory increases and decresases.
4. test case expectations have changed at some point. test_emmalloc_trim
was written with an assumption that emmalloc would acquire the whole
heap with sbrk() at startup.. but no longer does so. Adjusted test to
not assume this.
5. printing bare memory values in test doesn't work well in e.g. ubsan
test mode, so remove the direct value prints in the test, and just
assert() the expectations.

Fixes: #23343