Revert "Fix shlex splitting when launching parallel browser test runner on Windows." (#25229)
This broken the chrome tests in CI because `--use-gl=swiftshader` is no
longer part of `EMTEST_BROWSER` so `no_swiftshader` doesn't work and log
of the cubegeom tests are therefore failing.

It should be easy enough to re-land, but we need to consider how to make
the effects of `EMTEST_BROWSER_AUTO_CONFIG` visible in places like
`no_swiftshader`.

Reverts emscripten-core/emscripten#25222