Perform `pip install` as part of `bootstrap.py`
Also, install pip packages in local `out/python_deps` directory instead
of installing system wide.

This means we can consistently expect dev dependencies to be available
in test code without needing to include an opt out mechanism.

We were already doing this for `psutil`, but for `websockify` we made
it optional.

This means that only python scripts that explicitly add
`out/python_deps` to their python path will be able use
the packages, and in particular it means that the emscripten compiler
itself won't end up implicitly/accidentally depending on them.