Make -gsource-map independent of name sections and JS minification (#25238)
Previously, The `-gsource-map` flag (in addition to generating a source
map) caused a name section
to be generated in the wasm binary, and also suppressed minification of
the JS glue.
This PR makes -gsource-map independent of name sections and
minification, so it no longer does
either of those. This allows source maps to be used with fully-optimized
release builds, for use
cases such as symbolizing stack traces from production.
To get the old behavior, the `-g2` flag can be added to `-gsource-map`.

Fixes #20462
Fixes #24626