Fix `findCanvasEventTarget` finding default canvas with OFFSCREENCANVAS_SUPPORT to return the canvas object instead of the canvas target name (#25302)
When calling `findCanvasEventTarget`, with target name `canvas`, the
function returns the string key `canvas` instead of the actual canvas
object. From looking at the usages, it seems like it is expected for
this function to return the canvas object, not the target name string.

I found this issue when trying to call
[`wgpuInstanceCreateSurface`](https://github.com/google/dawn/blob/fe41c8b6d79ba421cb53430930cd037b40f8c185/third_party/emdawnwebgpu/pkg/webgpu/src/library_webgpu.js#L1979),
which is expecting the canvas object, but fails to find function
`getContext` on the returned string.

The function is returning the name because its returning the element of
the key array returned from `Object.keys(GL.offscreenCanvases)`, but
should be using `Object.values(GL.offscreenCanvases)` instead to return
the actual canvas object.