Implement support for reading argc and argv in MINIMAL_RUNTIME. (#25307)
Implement support for reading argc and argv in MINIMAL_RUNTIME. With the
help of the existing MAIN_READS_PARAMS machinery (thanks to
https://github.com/emscripten-core/emscripten/pull/9064), this is
only-pay-for-what-you-use.

Note that unlike the regular runtime that locates `argv[0]` to be the
script name, MINIMAL_RUNTIME will place the shell interpreter name in
`argv[0]` and the script name in `argv[1]`, to align with Node.js and
Python `process.argv` and `sys.argv` behavior.

In a browser run, regular runtime has a dummy `"./this.program"` in
`argv[0]`. Instead of a hardcoded mock value in `argv[0]`, the current
page URL (e.g. `"https://localhost:5932/foo/bar/baz.html"`) is located
in `argv[0]` instead, followed by the search query parameters in
`argv[1]` onwards.