Fix !DECLARE_ASM_MODULE_EXPORTS tests after #25298. NFC
In !DECLARE_ASM_MODULE_EXPORTS that wasm symbols are all programmatically
exported onto the Module object, and after #25298 that includes the
`wasmMemory` symbol.

This happens even for symbols that were not included in
EXPORTED_FUNCTIONS or EXPORTED_RUNTIME_SYMBOLS.  In many of these
tests `wasmMemory` was not being exported and so that
`unexportedRuntimeSymbol` helper was installing a property accessor on
the `Module` object, which did not all of the assignment within
`exportWasmSymbols`.

An alternative here would be to limit the list of symbols that
`exportWasmSymbols` assigns to the `Module` object but that would only
increase code size for very little benefit.