[empath-split] Tidy up error checking (#25380)
This puts all parser-related error checking routines in `parse_args` and
all other error checking routines, such as checking whether files exist,
in `check_errors`. The reason parser-related errors better be in
`parse_args` is we can use `parser.error` there, which prints the usage.

Also this checks for one more thing: whether `-o` or `--output` is
provided. This is forwarded to `wasm-split` and it is a mandatory option
for `wasm-split --multi-split`, but because `-o is required` error was
printed by not this script but `wasm-split`, it was printed later and
the message could be buried in warnings, which was confusing. We check
it here directly and error out when one is not provided.