Use ccache for emscripten library build (#25392)
Use ccache to cache the output of clang when building the emscripten
libraries.
When building on the main branch, save the ccache directory to the
CircleCI cache affter
building, using the clang version as the key. When building from other
branches (including PRs,
where the branch will appear as 'pull/12345') restore the cache before
building.