Remove ES6 constructs from file_packager output (#25476)
This allows the generated code to run on older engines.

Unlike the majority of the emscripten-generated code the file packager
output does not always get run through babel when targeting older
engines.

Specifically tests the standalone file packages code on the oldest
support node version.   This also required removing the use of the
`fs/Promises` packages which is not present in older versions of node.

Split out from #25459