Remove windows-specific console coloring
Windows 10 and above support ANSI coloring and there should be very few
emscripten users on older version of windows at this point (if any?).

The downside is that windows 7 will no longer have nice colored error
messages, but the upside is that we can cleanup and refactor this code
now (see #25495).