select: Allow registering cleanup method for the callback
This commit adds the registerCleanupFunc method for the callback
passed to the stream implementation. It allows the stream
implementation to optionally register a cleanup function that will be
invoked when select is no longer interested in the stream events. This
enables the stream implementation to perform additional cleanups such
as discarding the reference to the event callback.

Signed-off-by: Kohei Tokunaga <ktokunaga.mail@gmail.com>