Fix enabling of windows color terminal processing (#25529)
This was broken in #25502 and we only have limited testing of this
features. Specifically we only test that the ANSI codes are generated,
not that they are successfully processed.

The change here is not only to check for
ENABLE_VIRTUAL_TERMINAL_PROCESSING mode but to actually set this mode.

Replaces #25525