Add `MUTABLE_GLOBALS` to feature_matrix (#25532)
This is needed by dynamic linking since GOT entries are all imported as
mutable globals.

The net effect of this is that you cannot actually target the oldest
version of safari that we support (12) and with dynamically linked
programs. Specifying `-sLEGACY_VM_SUPPORT` will give you safari 13.1
instead of 12 if you also using `-sMAIN_MODULE`.