Only export selected data addresses on the Module object (#25576)
Just like for other types of symbols we don't want to export symbols on
the `Module` object itself if they user requested this.

Also, fix the memory profiler to not depend the external export of
`__heap_base` but instead the internal version (like it does for `_sbrk`
and other dependencies.