Mark CROSS_ORIGIN as requiring pthreads and dynamic execution (#25582)
This setting was recently added in #25581 and it only currently has en
effect on pthread-based programs.

It is also not compatible with `-sNO_DYNAMIC_EXECUTION` because it
depends on starting worker based on JS string.

These errors look like this:

```
$ emcc -sCROSS_ORIGIN -pthread -sDYNAMIC_EXECUTION=0 hello.c
emcc: error: CROSS_ORIGIN is not compatible with DYNAMIC_EXECUTION=0
```