Avoid using a new temp directory for each parallel test. NFC
It should be fine for all parallel tests to using the same TEMP_DIR.

The only limitation is on tests that use the canonical test directory
(e.g. EMCC_DEBUG).

It looks like we used to use a unique temp directory per process (See #6150).
However that we changes to one-directory-per-test in #18214, although
I don't remember why.

With this change we don't create a new temp directory except for tests
that are marked as `uses_canonical_tmp`