Remove tempfile usage from test_sanity.py. NFC
In both cases the tempfile usage was not necessary.

test_emconfig: In this case we can just create a new config file in
the CWD.  Because the test corrupts the main config file there is no
risk that the test is accidentally finding the config file via CWD alone
so there is no need to change directory into an external temp dir.

test_firstrun: We an just create the fake `bin` files in a subdirectory
of the current directory where the test runs.  No need for an external
temp dir here.