[memoryprofiler] Add CSS class to parent div (#25595)
Add a CSS class to the main parent div of the memory profiler UI.

Currently, the profiler's container div is created with no ID or class,
making it impossible to target with CSS. This is problematic for
developers (like in the original issue) who run their application in
full-screen, as the profiler UI gets hidden with no way to adjust its
`z-index`.

This change adds the class `emscripten-memory-profiler-container` to the
div upon its creation in `library_memoryprofiler.js`.

This allows developers to easily style the profiler. For example:

```
<style>
  .emscripten-memory-profiler-container {
    z-index: 9999;
  }
</style>
```

Fixes: #20115