[empath-split] Support multi-paths modules (#25577)
This changes the format of the user input "paths" file so that multiple
paths can be split into a single module. The new paths file structure is
now very similar to wasm-split's manifest file, with functions replaced
with paths. For example, the format will be like
```
module1
path/to/a
path/to/b

module2
path/to/c
```
Where `module1` and `module2` are module names.