Improve Firefox browser process tracking (#25644)
Improve Firefox browser process tracking by using the -wait-for-browser
cmdline arg, and waiting longer for process startup.

Unfortunately `-wait-for-browser` is not enough to get rid of the
process delta polling, since the process handle returned by
`subprocess.Popen()` cannot be used to cascade the spawned Firefox
windows.

Waiting for 2 seconds for Firefox startup was not enough, so make the
wait time more conservative.

Add the detected processes to the list rather than replace, so that all
process handles are retained.

It may be possible to remove the use of
`launch_browser_harness_with_proc_snapshot_workaround()` for headless
Firefox runs, but more testing will be needed to verify that.