Don't require __wasm_apply_data_relocs. NFC
This symbol is optionally created by the linker and optionally used
in preamble.js and libdylink.js.

However this requirement here was forcing it to be exported.  The only
reason I think this didn't cause any issues is that its not really
possible to create a main module with zero data relocations today.