[SPLIT_MODULE] Fix imports into deferred module when loading (#25621)
Fix an error when attempted to reserve a new memory space in the
deferred module when building with MEMORY64 + SPLIT_MODULE.
```
TypeError: Cannot convert <func ptr> to a BigInt
```

In a mem64 build, we call applySignatureConversions on `malloc()` in the
wasmExports. `applySignatureConversions()` wraps the exports and
converts the return value to a javascript number. The wasmExports is
passed into the deferred module during instantiation so when we call a
malloc in the deferred module, we'd be calling the wrapped function,
hitting into this error.

Furthermore, we are making a direct call when calling a malloc, so
strictly speaking, we should not need to go through the JS trampoline.