Sort members of EmccOptions. NFC
Also, use class variables instead of instance fields. This doesn't
change the semantics (at least not the way we use this class) and it
saves a little bit of visual noise.

The other nice thing this would allow if we ever wanted to to it is to
compare the current option with its default (by comparing `self.foo`
with `__class__.foo`, the default value).