Fix issue with transitive/recursive building of PIC libraries (#25707)
This issue was introduces in #25522 when `-sRELOCATABLE` was removed
from `get_base_cflags.

This is a somewhat tricky to reproduce / understand issue with
transitive library building:

1. User attempts to compile program using SDL2 + SDL2_GFX.
2. `emcc` attempts to builds all needed libraries in reverse dependency
order. And ends up with the follows sequence:
  1. libsdl2-pic
  2. libsdl2_gfx-pic
3. `emcc` builds libsdl2-pic
4. `emcc` builds libsdl2_gfx-pic. The command that it constructs when
building each source file looks like `emcc -fPIC gfx.c -sUSE_SDL=2`.
However because neither `-sMAIN_MODULE` nor `-sRELOCATABLE` were
specified this `emcc` command will not try to build the non-PIC version
of libsdl2. At this point the compiler will crash with:

`AssertionError: attempt to lock the cache while a parent process is
holding the lock (sysroot/lib/wasm32-emscripten/libSDL2.a)`

This issue is hard to reproduce because if libsdl2 is already built it
will not trigger the issue.