Make command line options global. NFC (#25708)
Refactor only. This should allows us to start removing some of the
internal settings that are currently used to communicate command line
flags from part of the python codebase to another. We should only need
to add internal settings when a setting needs to be available in the JS
compiler.