Add `--buffer` option to test runner. NFC (#25737)
This is standard option in the default python unintest runner.

It now works for both the normal serial runner as well as the parallel
runner. In the case of the parallel runner we not inherit from the
standard `unittest.TestResult` so that we can inherit the the
stdout/stderr buffering methods it implements.

In the future we may want to enable this by default.  

This could especially useful the parallel runner where we have issue of
interleaving the stdout and stderr of different tests is confusing.