Remove a bunch of redundant code from parallel_testsuite.py
All this infrastructure with the BufferedTestBase was added to work
around that fact that we were previously using using a fake TestResult
class so that is could be serialized.

However in #25737 I converted to using TestResult as the superclass of
BufferedParallelTestResult.  There were only two members that needed to
me removed to make it work.

With that change landed we can now remove a log of the other code here.