Add `list()` function to `HandleAllocator` for listing valid ids
Since it is possible to build such a list by manually iterating
over the `allocated` array, I think it is better to provide a function
that does just that.
Accessing the `allocated` member from outside the class should be
avoided as it is more an implementation detail than a part of the
class interface.