[AUDIO_WORKET] Add support for setting the render quantum size (#25747)
The `renderSizeHint` is now available in Chrome Canary
(`--enable-features=WebAudioConfigurableRenderQuantum`) so this can
finally be tested. Most of the work was already done in preparation, so
this required few changes.

Test with:
```
test/runner interactive.test_audio_worklet_params_mixing
```
Which has been extended to request a large `2048` sample audio frame
(which in an older browser will fallback to the default 128 samples).

Note for me: look at why and since how long we can bust out of the stack
without asserting. The stackAlloc() no longer appears to trip when
requesting more bytes than `wwParams.stackSize`.