Remove a bunch of redundant code from parallel_testsuite.py (#25739)
All this infrastructure with the BufferedTestBase was added to work
around that fact that we were previously using using a fake TestResult
class so that is could be serialized.

However in #25737 I converted to using TestResult as the superclass of
BufferedParallelTestResult. There were only two members that needed to
me removed to make it work. The `TestResult` class stores all exceptions
as strings so there is also no need to the Fake exceptions stuff
anymore.