Simplify error handling in parallel_testsuite
The `test(result)` call already contains a try/catch for landing errors
and skipping internally so the extra exception blocks here were not
doing anything.

What is more, if they ever did execute they would fail because the
second argument to `addError` and `addSkip` is not an exception
object but and exception info triple.