Fix memory leak on cross thread events (#25783)
I believe that we started leaking the event data for cross-thread html5
events in #20370.

Prior to that change we use `_emscripten_dispatch_to_thread` which took
care of free'ing the event data once it was consumed on the target
thread.

After this change we copy make copy of the event strut in
`_emscripten_run_callback_on_thread` and then free it once the callback
is run.

This change also avoids an extra malloc per event, and cleans up
libhtml5.js a little.