[fetch] Support streaming fetch requests. (#25383)
Add a new `FETCH_STREAMING` setting that enables using the DOM fetch API
to stream data when paired with `EMSCRIPTEN_FETCH_STREAM_DATA`. When
`EMSCRIPTEN_FETCH_STREAM_DATA` is not used we will default to a regular
XMLHttpRequest. To keep the emscripten Fetch.js code the same I've
implemented a polyfill of XMLHttpRequest using Fetch. To support
streaming I wired up the fetch code to use the old onprogress code that
old versions of Firefox supported.

Most of the current API is supported with some notable exceptions:
 - synchronous requests
 - overriding the mime type

I also changed a few of the tests to support both sync and async so I
could test them with the fetch backend.