[dylink] Fix rpath calculation in nested dependencies (#24234)
This fixes a bug in rpath calculation in nested dependencies.

Basically, it fixes a case when a relative path is passed to the
`loadDynamicLibrary`.

Think of the following scenario:

0. libhello.wasm depends on libhello_dep.so depends on
libhello_nested_dep.so
1. [main.c] calls dlopen('/usr/lib/libhello.wasm')
2. [dynlink.c] calls `loadDynamicLibrary('/usr/lib/libhello.wasm')`
(_dlopen_js ==> dlopenInternal ==> loadDynamicLibrary)
3. [libdylink.js] inside `loadDynamicLibrary('/usr/lib/libhello.wasm')`,
it loads the dependency `libhello_dep.so`. When locating
`libhello_dep.so`, it uses `parentLibPath: /usr/lib/libhello.wasm` it
locates the library correctly from `/usr/lib/libhello_dep.so`.
5. [libdylink.js] Now we call `loadDynamicLibrary('libhello_dep.so')`,
and it loads its dependency 'libhello_nested_dep.so' first. In this case
it uses `parentLibPath: libhello_dep.so` so it fails to locate the
library.